moved buildRLPMessage(), buildIDTag() to hashing.go; added decodeRLPMessage()
