+++ aeternity/transactions.go
@@ -2,8 +2,6 @@ package aeternity
 
 import (
 	"fmt"
-
-	"github.com/aeternity/aepp-sdk-go/rlp"
 )
 
 // SignEncodeTx sign and encode a transaction
@@ -24,28 +22,6 @@ func SignEncodeTx(kp *Account, txRaw []byte) (signedEncodedTx, signedEncodedTxHa
 	signedEncodedTxHash = encode(PrefixTransactionHash, rlpTxHashRaw)
 	// encode the signature
 	signature = encode(PrefixSignature, sigRaw)
-	return
-}
-
-func buildRLPMessage(tag uint, version uint, fields ...interface{}) (rlpRawMsg []byte, err error) {
-	// create a message of the transaction and signature
-	data := []interface{}{tag, version}
-	data = append(data, fields...)
-	// fmt.Printf("TX %#v\n\n", data)
-	// encode the message using rlp
-	rlpRawMsg, err = rlp.EncodeToBytes(data)
-	// fmt.Printf("ENCODED %#v\n\n", data)
-	return
-}
-
-// buildIDTag assemble an id() object
-// see https://github.com/aeternity/protocol/blob/epoch-v0.22.0/serializations.md#the-id-type
-func buildIDTag(IDTag uint8, encodedHash string) (v []uint8, err error) {
-	raw, err := decode(encodedHash)
-	v = []uint8{IDTag}
-	for _, x := range raw {
-		v = append(v, uint8(x))
-	}
 	return
 }
